Rename variables -> variable
