+++ pyccel/ast/mathext.py
@@ -7,7 +7,7 @@ import math
 from sympy import Function
 
 from pyccel.ast.basic     import PyccelAstNode
-from pyccel.ast.variables import Constant
+from pyccel.ast.variable  import Constant
 from pyccel.ast.datatypes import (NativeInteger, NativeBool, NativeReal,
                                   default_precision)
