+++ pyccel/parser/syntax/headers.py
@@ -18,7 +18,7 @@ from pyccel.ast.headers   import FunctionHeader, ClassHeader, MethodHeader, Vari
 from pyccel.ast.headers   import MetaVariable , UnionType, InterfaceHeader
 from pyccel.ast.headers   import construct_macro, MacroFunction, MacroVariable
 from pyccel.ast.core      import ValuedArgument
-from pyccel.ast.variables import DottedName
+from pyccel.ast.variable  import DottedName
 from pyccel.ast.datatypes import dtype_and_precision_registry as dtype_registry, default_precision
 from pyccel.ast.literals  import LiteralString
 from pyccel.errors.errors import Errors
